Separate the various concepts of options (#26)
The prior code used `task_options` to mean several
distinct things, which get layered. Make it more
obvious what's going on by separating the names
a bit.

Additionally, the layering of options for `Tasks`
were certainly not being respected, because downstream
classes were reaching into the raw stores and were
expected to compute the layering for themselves. At a
minimum, this is fragile. Instead, hide the
implementation, so we can rely on it.