DE-7776 Upgrade sqlalchemy to 1.4 and fix deprecation warnings (#266)
* Upgrade sqlalchemy to >1.4.0

* Fix simple deprecation warnings

* Bump isort in pre-commit

* Fix warnings related to ORDER BY and SELECT

RemovedIn20Warning: ORDER BY columns added implicitly due to DISTINCT is
deprecated and will be removed in SQLAlchemy 2.0.  SELECT statements
with DISTINCT should be written to explicitly include the appropriate
columns in the columns clause (Background on SQLAlchemy 2.0 at:
https://sqlalche.me/e/b8d9)

RemovedIn20Warning: The legacy calling style of select() is deprecated
and will be remo ved in SQLAlchemy 2.0.  Please use the new calling
style described at select(). (Background on SQLAlchemy 2.0 at:
https://sqlalche.me/e/b8d9)

* Fix Query.from_self deprecation warnings

RemovedIn20Warning: The Query.from_self() method is considered legacy as
of the 1.x s eries of SQLAlchemy and will be removed in 2.0. The new
approach is to use the orm.aliased() construct in conjunction with a
subquery.  See the section "Selecting from the query itself as a
subquery" in the 2.0 migration notes for an example. (Background on
SQLAlchemy 2.0 at: https://sqlalche.me/e/b8d9)

* Elevate sqlalchemy 2.0 warnings to errors

* Use file display for file-path logs

* Revert display change